Refactor handling of 'site' field in multiple resources and state functions
