+++ internal/endpoints/accountgroups/accountgroups_resource.go
@@ -10,6 +10,8 @@ import (
 	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
 	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/client"
 	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/jamfprivileges"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/state"
 	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/waitfor"
 
 	util "github.com/deploymenttheory/terraform-provider-jamfpro/internal/helpers/type_assertion"
@@ -102,7 +104,7 @@ func ResourceJamfProAccountGroups() *schema.Resource {
 				Computed:    true,
 				Elem: &schema.Schema{
 					Type:         schema.TypeString,
-					ValidateFunc: common.ValidateJSSObjectsPrivileges,
+					ValidateFunc: jamfprivileges.ValidateJSSObjectsPrivileges,
 				},
 			},
 			"jss_settings_privileges": {
@@ -112,7 +114,7 @@ func ResourceJamfProAccountGroups() *schema.Resource {
 				Computed:    true,
 				Elem: &schema.Schema{
 					Type:         schema.TypeString,
-					ValidateFunc: common.ValidateJSSSettingsPrivileges,
+					ValidateFunc: jamfprivileges.ValidateJSSSettingsPrivileges,
 				},
 			},
 			"jss_actions_privileges": {
@@ -122,7 +124,7 @@ func ResourceJamfProAccountGroups() *schema.Resource {
 				Computed:    true,
 				Elem: &schema.Schema{
 					Type:         schema.TypeString,
-					ValidateFunc: common.ValidateJSSActionsPrivileges,
+					ValidateFunc: jamfprivileges.ValidateJSSActionsPrivileges,
 				},
 			},
 			"casper_admin_privileges": {
@@ -132,7 +134,7 @@ func ResourceJamfProAccountGroups() *schema.Resource {
 				Computed:    true,
 				Elem: &schema.Schema{
 					Type:         schema.TypeString,
-					ValidateFunc: common.ValidateCasperAdminPrivileges,
+					ValidateFunc: jamfprivileges.ValidateCasperAdminPrivileges,
 				},
 			},
 			"members": {
@@ -266,7 +268,7 @@ func ResourceJamfProAccountGroupRead(ctx context.Context, d *schema.ResourceData
 
 	if err != nil {
 		// Handle not found error or other errors
-		return common.HandleResourceNotFoundError(err, d)
+		return state.HandleResourceNotFoundError(err, d)
 	}
 
 	// Update the Terraform state with the fetched data from the resource
