+++ internal/endpoints/accounts/accounts_object.go
@@ -7,7 +7,7 @@ import (
 	"log"
 
 	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
-	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/constructobject"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -34,10 +34,10 @@ func constructJamfProAccount(d *schema.ResourceData) (*jamfpro.ResourceAccount, 
 
 	// Handle Site
 	if v, ok := d.GetOk("site"); ok {
-		account.Site = common.ConstructSharedResourceSite(v.([]interface{}))
+		account.Site = constructobject.ConstructSharedResourceSite(v.([]interface{}))
 	} else {
 		// Set default values if 'site' data is not provided
-		account.Site = common.ConstructSharedResourceSite([]interface{}{})
+		account.Site = constructobject.ConstructSharedResourceSite([]interface{}{})
 	}
 	// Handle Privileges
 	account.Privileges = constructAccountSubsetPrivileges(d)
