+++ internal/endpoints/advancedmobiledevicesearches/advancedmobiledevicesearches_object.go
@@ -7,7 +7,7 @@ import (
 	"log"
 
 	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
-	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/constructobject"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -55,10 +55,10 @@ func constructJamfProAdvancedMobileDeviceSearch(d *schema.ResourceData) (*jamfpr
 
 	// Handle Site
 	if v, ok := d.GetOk("site"); ok {
-		search.Site = common.ConstructSharedResourceSite(v.([]interface{}))
+		search.Site = constructobject.ConstructSharedResourceSite(v.([]interface{}))
 	} else {
 		// Set default values if 'site' data is not provided
-		search.Site = common.ConstructSharedResourceSite([]interface{}{})
+		search.Site = constructobject.ConstructSharedResourceSite([]interface{}{})
 	}
 
 	// Serialize and pretty-print the Advanced Mobile Device Search object as XML for logging
