+++ internal/endpoints/advancedmobiledevicesearches/advancedmobiledevicesearches_resource.go
@@ -10,6 +10,7 @@ import (
 	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
 	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/client"
 	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/state"
 	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/waitfor"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
@@ -218,7 +219,7 @@ func ResourceJamfProAdvancedMobileDeviceSearchRead(ctx context.Context, d *schem
 
 	if err != nil {
 		// Handle not found error or other errors
-		return common.HandleResourceNotFoundError(err, d)
+		return state.HandleResourceNotFoundError(err, d)
 	}
 
 	// Update the Terraform state with the fetched data from the resource
