+++ internal/endpoints/advancedusersearches/advancedusersearches_object.go
@@ -7,7 +7,7 @@ import (
 	"log"
 
 	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
-	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/constructobject"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -51,10 +51,10 @@ func constructJamfProAdvancedUserSearch(d *schema.ResourceData) (*jamfpro.Resour
 
 	// Handle Site
 	if v, ok := d.GetOk("site"); ok {
-		search.Site = common.ConstructSharedResourceSite(v.([]interface{}))
+		search.Site = constructobject.ConstructSharedResourceSite(v.([]interface{}))
 	} else {
 		// Set default values if 'site' data is not provided
-		search.Site = common.ConstructSharedResourceSite([]interface{}{})
+		search.Site = constructobject.ConstructSharedResourceSite([]interface{}{})
 	}
 
 	// Serialize and pretty-print the Advanced User Search object as XML for logging
