+++ internal/endpoints/computergroups/computergroups_object.go
@@ -7,7 +7,7 @@ import (
 	"log"
 
 	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
-	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/constructobject"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -20,10 +20,10 @@ func constructJamfProComputerGroup(d *schema.ResourceData) (*jamfpro.ResourceCom
 
 	// Handle Site
 	if v, ok := d.GetOk("site"); ok {
-		group.Site = common.ConstructSharedResourceSite(v.([]interface{}))
+		group.Site = constructobject.ConstructSharedResourceSite(v.([]interface{}))
 	} else {
 		// Set default values if 'site' data is not provided
-		group.Site = common.ConstructSharedResourceSite([]interface{}{})
+		group.Site = constructobject.ConstructSharedResourceSite([]interface{}{})
 	}
 	// Handle "criteria" field
 	if v, ok := d.GetOk("criteria"); ok {
