+++ internal/endpoints/usergroups/usergroups_object.go
@@ -8,7 +8,7 @@ import (
 	"strconv"
 
 	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
-	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/constructobject"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -22,10 +22,10 @@ func constructJamfProUserGroup(d *schema.ResourceData) (*jamfpro.ResourceUserGro
 
 	// Handle Site
 	if v, ok := d.GetOk("site"); ok {
-		userGroup.Site = common.ConstructSharedResourceSite(v.([]interface{}))
+		userGroup.Site = constructobject.ConstructSharedResourceSite(v.([]interface{}))
 	} else {
 		// Set default values if 'site' data is not provided
-		userGroup.Site = common.ConstructSharedResourceSite([]interface{}{})
+		userGroup.Site = constructobject.ConstructSharedResourceSite([]interface{}{})
 	}
 
 	criteria := d.Get("criteria").([]interface{})
